requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.10.0
bootstrap4==0.1.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.2.7
django-bootstrap4==3.0.1
django-embed-video==1.4.0
django-environ==0.7.0
environ==1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
pytz==2021.1
requests==2.26.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
style==1.1.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
update==0.0.1
urllib3==1.26.7
zope.interface==5.4.0
